# Add SegmentFromPoint ROS Node