--- a/ada_feeding_perception/install.sh
+++ b/ada_feeding_perception/install.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+pip install git+https://github.com/facebookresearch/segment-anything.git
+pip install -r requirements.txt