--- a/ada_feeding_perception/requirements.txt
+++ b/ada_feeding_perception/requirements.txt
@@ -1,0 +1,5 @@
+numpy
+torch
+matplotlib
+opencv-python
+scikit-image