--- a/ada_feeding_perception/requirements.txt
+++ b/ada_feeding_perception/requirements.txt
@@ -1,6 +1,7 @@
 numpy
-torch
 matplotlib
 opencv-python
 scikit-image
 scipy
+torch
+torchvision