@amalnanavati 2023-06-08
Ah no worries, that script is very resource intensive. I think it'll only be useful on GPU machines.